+++ apis/r/NEWS.md
@@ -8,7 +8,8 @@
 
 ## Deprecated
 
-- `SOMANDArrayBase$set_data_type()` is marked for deprecation begining with tiledbsoma 2.1.0
+- Mark `SOMANDArrayBase$set_data_type()` for deprecation. ([#4279](https://github.com/single-cell-data/TileDB-SOMA/pull/4279))
+- Mark setting `domain` to `NULL` in `SOMADataFrame$create()` for deprecation. ([#4274](https://github.com/single-cell-data/TileDB-SOMA/pull/4274))
 
 ## Removed